requirements.txt
python-telegram-bot==13.7
spotipy==2.19.0
youtube-dl==2021.12.17
instaloader==4.9.5
requests==2.28.1
pydub==0.25.1
ffmpeg-python==0.2.0
python-dotenv==0.21.0